confirmation for deleting classes

+++ www/js/controllers/classListCtrl.js
@@ -2,35 +2,67 @@
 
 angular.module('hwo').controller('ClassListCtrl', ClassListCtrl);
 
-function ClassListCtrl($scope, $state, modifyClass, Class, Assignment) {
+function ClassListCtrl($scope, $state, $ionicPopup, $q, modifyClass, Class, Assignment) {
     loadClasses();
+
+    var editPromise = $q.when();
     
     $scope.assignments = assignments;
     $scope.delete = deleteClass;
     $scope.move = move;
     $scope.toggleEditing = toggleEditing;
     $scope.modals = {
-        edit: modifyClass.modals.edit,
+        edit: function(klass) {
+            editPromise.then(function(){
+                return modifyClass.modals.edit(klass);
+            });
+        },
         insert: function () {
             modifyClass.modals.insert().then(loadClasses);
         }
+
     };
     
     function assignments(klass) {
         $state.go('assignments', { classId: klass.id });
     }
     
-    function deleteClass(klass) {
-        var classes = $scope.classes;
-        var length = classes.length;
-        for (var i = klass.index + 1; i < length; i++) {
-            var other = classes[i];
-            other.index--;
-        }
+    function deleteClass($event, klass) {
+        editPromise = $ionicPopup.confirm({
+            title: 'HWTracker',
+            template: 'Are you sure you want to delete the class?'
+        }).then(function(res) {
+            if(res){
+                var classes = $scope.classes;
+                var length = classes.length;
+                for (var i = klass.index + 1; i < length; i++) {
+                    var other = classes[i];
+                    other.index--;
+                }
         
-        klass.delete();
-        classes.splice(klass.index, 1);
+                klass.delete();
+                classes.splice(klass.index, 1);
+            }
+            editPromise = $q.when();
+            return $q.reject();
+        });
     }
+
+//    function deleteClass($event, klass) {
+//        if(confirm("Are you sure you want to delete the class?")){
+//            var classes = $scope.classes;
+//            var length = classes.length;
+//            for (var i = klass.index + 1; i < length; i++) {
+//                var other = classes[i];
+//                other.index--;
+//            }
+//            
+//            klass.delete();
+//            classes.splice(klass.index, 1);
+//        } else {
+//            $event.preventDefault();
+//        }
+//    }
     
     function move(klass, fromIndex, toIndex) {
         if (fromIndex !== toIndex) {
